feat(plugin-word-cloud): change how wordcloud is exported
