+++ plugins/plugin-chart-word-cloud/src/legacyPlugin/index.ts
@@ -1,6 +1,6 @@
 import { t } from '@superset-ui/translation';
 import { ChartMetadata, ChartPlugin } from '@superset-ui/chart';
-import buildQuery from '../buildQuery';
+import buildQuery from '../plugin/buildQuery';
 import transformProps from './transformProps';
 import thumbnail from '../images/thumbnail.png';
 import { LegacyWordCloudFormData } from './types';
@@ -13,7 +13,7 @@ const metadata = new ChartMetadata({
   useLegacyApi: true,
 });
 
-export default class WordCloudChartPlugin extends ChartPlugin<LegacyWordCloudFormData> {
+export default class LegacyWordCloudChartPlugin extends ChartPlugin<LegacyWordCloudFormData> {
   constructor() {
     super({
       buildQuery,
