+++ plugins/plugin-chart-word-cloud/src/plugin/buildQuery.ts
@@ -1,5 +1,5 @@
 import { buildQueryContext } from '@superset-ui/query';
-import { WordCloudFormData } from './types';
+import { WordCloudFormData } from '../types';
 
 export default function buildQuery(formData: WordCloudFormData) {
   // Set the single QueryObject's groupby field with series in formData
