+++ plugins/plugin-chart-word-cloud/src/plugin/transformProps.ts
@@ -1,6 +1,6 @@
 import { ChartProps } from '@superset-ui/chart';
-import { WordCloudProps } from './chart/WordCloud';
-import { WordCloudFormData } from './types';
+import { WordCloudProps } from '../chart/WordCloud';
+import { WordCloudFormData } from '../types';
 
 export default function transformProps(chartProps: ChartProps): WordCloudProps {
   const { width, height, formData, queryData } = chartProps;
