+++ plugins/plugin-chart-word-cloud/test/index.test.ts
@@ -1,0 +1,10 @@
+import { WordCloudChartPlugin, LegacyWordCloudChartPlugin } from '../src';
+
+describe('plugin-chart-word-cloud', () => {
+  it('exports WordCloudChartPlugin', () => {
+    expect(WordCloudChartPlugin).toBeDefined();
+  });
+  it('exports LegacyWordCloudChartPlugin', () => {
+    expect(LegacyWordCloudChartPlugin).toBeDefined();
+  });
+});
